taydialog/dialogflow_ros/dialogflow_ros/requirements.txt
PyAudio
google-cloud-speech
six
google-api-python-client
dialogflow
pyyaml
protobuf-to-dict
